Refactor GitHubApp into GitHub, and ditch SourceProvider controller
